Removed a completely unnecessary cast to byte in the constructor.

+++ src/free/chess/Square.java
@@ -61,8 +61,8 @@ public class Square{
    */
   
   private Square(int file, int rank){
-    this.file = (byte)file;
-    this.rank = (byte)rank;
+    this.file = file;
+    this.rank = rank;
   }
